refactor(engine): extract core/ + dissolve envelope/
Restructure engine module layout to make the phase ↔ kernel split
explicit. Phases (extraction/detection/deduplication/redaction/
validation) depend on `core/` only; the orchestrator depends on
`core/` plus every phase module. `envelope/` disappears as a
top-level concept.

- New `core/` module: contains the phase contract surface and the
  per-run state phases see through it.
  - phase.rs       (Phase<M>, PhaseContext, PhaseInfo, ModalityKind)
  - target.rs      (PhaseTarget, ValueAt trait + impls, SharedHandle)
  - plan.rs        (Plan)
  - context.rs     (RunContext, moved from pipeline/orchestrator)
  - shared.rs      (SharedData, moved from envelope/)
  - policy_store.rs (PolicyStore + Decision, moved from envelope/)

- `pipeline/` keeps only execution + lifecycle + run history:
  - engine.rs      (renamed from default.rs: Engine + EngineInput + EngineOutput)
  - run.rs         (Pipeline)
  - orchestrator.rs (Orchestrator + DocumentPipeline<M>)
  - envelope.rs    (merged from envelope/mod.rs + envelope/any.rs)
  - config/, runs/ unchanged

- `envelope/` directory deleted. `envelope/accessors.rs` was dead
  after the PhaseTarget refactor and gets dropped; `envelope/
  value_at.rs` folds into `core/target.rs` next to its impls.

Phases keep working unchanged via a temporary `pipeline::*`
re-export of the core types; that's a clean follow-up to migrate
imports site-by-site.
