+++ crates/nvisy-engine/src/core/shared.rs
